# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "面向个人房东的租房管理后端系统，支持房源管理、租客管理、水电气收费、租金账单等功能。",
        "title": "租房管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/bills": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取账单列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账单管理"
                ],
                "summary": "获取账单列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "房间ID",
                        "name": "room_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账单月份",
                        "name": "bill_month",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "生成账单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账单管理"
                ],
                "summary": "生成账单",
                "parameters": [
                    {
                        "description": "账单信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_service.CreateBillRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/bills/monthly-statistics": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取年度月度统计",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账单管理"
                ],
                "summary": "获取月度统计",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 2024,
                        "description": "年份",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/bills/statistics": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取账单统计",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账单管理"
                ],
                "summary": "获取账单统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "开始日期",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/bills/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取账单详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账单管理"
                ],
                "summary": "获取账单详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/bills/{id}/pay": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "标记账单已支付",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账单管理"
                ],
                "summary": "支付账单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "支付信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_service.PayBillRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/fee-rates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取水电气费率配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统配置"
                ],
                "summary": "获取费率配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新水电气费率配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统配置"
                ],
                "summary": "更新费率配置",
                "parameters": [
                    {
                        "description": "费率信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_service.UpdateFeeRateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/meter-readings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取抄表记录列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "抄表管理"
                ],
                "summary": "获取抄表记录列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "房间ID",
                        "name": "room_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始日期",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "录入水电气表读数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "抄表管理"
                ],
                "summary": "创建抄表记录",
                "parameters": [
                    {
                        "description": "抄表信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_service.CreateMeterReadingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "rental-management_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "rental-management_internal_service.CreateBillRequest": {
            "type": "object",
            "required": [
                "bill_month",
                "room_id"
            ],
            "properties": {
                "bill_month": {
                    "description": "YYYY-MM",
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "electricity_fee": {
                    "type": "number"
                },
                "gas_fee": {
                    "type": "number"
                },
                "rent_fee": {
                    "type": "number"
                },
                "room_id": {
                    "type": "integer"
                },
                "tenant_id": {
                    "type": "integer"
                },
                "water_fee": {
                    "type": "number"
                }
            }
        },
...
        "rental-management_internal_service.CreateMeterReadingRequest": {
            "type": "object",
            "required": [
                "reading_date",
                "room_id"
            ],
            "properties": {
                "electricity_reading": {
                    "type": "number"
                },
                "gas_reading": {
                    "type": "number"
                },
                "reading_date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "room_id": {
                    "type": "integer"
                },
                "water_reading": {
                    "type": "number"
                }
            }
        },
...
        "rental-management_internal_service.PayBillRequest": {
            "type": "object",
            "required": [
                "amount",
                "payment_method"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "note": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "integer"
                }
            }
        },
...
        "rental-management_internal_service.UpdateFeeRateRequest": {
            "type": "object",
            "properties": {
                "electricity_rate": {
                    "type": "number"
                },
                "gas_rate": {
                    "type": "number"
                },
                "water_rate": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}